refactor: 원래 OuputView를 ConsoleOuputView로 변경하고 OuputView를 interface로 만들어 ConsoleOuputView가 상속하여 변화에 대비할 수 있게 변경
