+++ src/main/kotlin/racingcar/view/output/OutputView.kt
@@ -1,0 +1,14 @@
+package racingcar.view.output
+
+interface OutputView {
+    fun enterCarNames()
+
+    fun enterNumberOfAttempts()
+
+    fun printExecutionResults(
+        attempts: Int,
+        intermediateResults: List<String>,
+    )
+
+    fun printLastWinner(winners: List<String>)
+}
